sorter(ticdc): rename files for better readability
